+++ src/numbers/xdr_large_int.ts
@@ -165,9 +165,6 @@ export class XdrLargeInt {
     this._sizeCheck(128);
 
     const v = this.int.toBigInt();
-    if (BigInt.asIntN(128, v) !== v) {
-      throw RangeError(`value too large for i128: ${v}`);
-    }
     const hi64 = BigInt.asIntN(64, v >> 64n); // encode top 64 w/ sign bit
     const lo64 = BigInt.asUintN(64, v); // grab btm 64, encode sign